#77 Moves txclient into the db pkg.
